+++ documentation-website/src/client/components/Content.js
@@ -1,5 +1,19 @@
 import React from "react";
+import { Focus } from "@curi/react";
 
 import "../scss/content.scss";
 
-export default ({ children }) => <div className="content">{children}</div>;
+export default ({ children }) => (
+  <Focus>
+    {ref => (
+      <div
+        className="content"
+        tabIndex={-1}
+        ref={ref}
+        style={{ outline: "none" }}
+      >
+        {children}
+      </div>
+    )}
+  </Focus>
+);